+++ back/.gitignore
@@ -15,6 +15,7 @@
 # misc
 .DS_Store
 .env
+.env.test
 .env.local
 .env.development.local
 .env.test.local
