+++ back/package.json
@@ -8,7 +8,8 @@
   "dependencies": {
     "body-parser": "^1.19.0",
     "dotenv": "^8.0.0",
-    "express": "^4.17.0"
+    "express": "^4.17.0",
+    "http-status-codes": "^1.3.2"
   },
   "devDependencies": {
     "chai": "^4.2.0",
@@ -25,10 +26,10 @@
   "scripts": {
     "coverage": "nyc mocha --exit && nyc report --reporter=text-lcov > ../.lcov/back.lcov",
     "coveralls": "lcov-result-merger '../.lcov/*.lcov' | coveralls",
-    "lint": "eslint src --ext .js,.jsx",
+    "lint": "eslint src --ext .js && eslint test --ext .js",
     "start": "cd src && node index",
     "test": "mocha",
-    "test-unwatch": "mocha --exit"
+    "test-unwatch": "yarn test --exit"
   },
   "eslintConfig": {}
 }
