+++ back/src/app/routes.js
@@ -1,9 +1,8 @@
+const { OK } = require('http-status-codes');
 const UserController = require('../model/user.controller');
 
 module.exports = (app) => {
-  app.get('/', (_, res) => {
-    res.status(200).json({ message: 'Ok, auto deploy working' });
-  });
+  app.get('/', (_, res) => res.status(OK).json({ message: 'OK, working...' }));
 
   app.route('/users')
     .get(UserController.list);
